division of the scroll into different components
